+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/AlbumEditPanel.java
@@ -30,7 +30,7 @@ public final class AlbumEditPanel extends EntityEditPanel {
             .add(true)
             .edit(true);
     createTextField(Album.TITLE);
-    setComponent(Album.COVER, new CoverArtPanel(editModel().value(Album.COVER)));
+    component(Album.COVER).set(new CoverArtPanel(editModel().value(Album.COVER)));
 
     JPanel centerPanel = borderLayoutPanel()
             .westComponent(borderLayoutPanel()